compiler: Phase 4 — target-aware JIT optimizer + fast-math; diagnose GEMM perf
Investigate the CPU JIT GEMM perf lever. Measured: tessera_jit's
linalg->loops->LLVM GEMM runs ~2.2 GFLOP/s (256/512 cubed), ~50-110x off
numpy/Accelerate — the ConvertLinalgToLoops body is naive scalar.

Two cheap levers, both measured insufficient but kept because they are correct:
- Host-detected TargetMachine into the transformer (was nullptr -> no NEON cost
  model) + -O3, so the LLVM optimizer is target-aware.
- Stamp fastmath<fast> on float arith ops after linalg->loops (a float reduction
  acc+=a*b won't auto-vectorize without reassoc). 'fast' matches Tessera's
  documented fast-math GEMM contract.

Neither moves the GEMM (LLVM's loop vectorizer won't crack the reduction from
this IR shape). The real lever is an MLIR linalg->vector tiling+vectorization
pipeline (a focused multi-step effort); documented in COMPILER_AUDIT with scope
honesty (won't match BLAS; single-GEMM hot path already uses Accelerate).

Also fixes two stale tests the earlier f64 wiring invalidated (they asserted f64
is rejected; now assert f64 executes + use int32 as the unsupported-dtype probe).

3062 JIT/ops/fusion tests green, 0 failures; drift clean.

+++ tests/unit/test_production_jit_add.py
@@ -59,9 +59,9 @@ def test_jit_add_is_destination_passing_writes_into_fresh_output():
 @pytest.mark.parametrize(
     "a, b",
     [
-        # f64 — dtype outside the f32/f16/bf16 boundary table (native on M1 but
-        # not yet wired; f16 is now supported, see test below).
-        (np.ones((4, 4), np.float64), np.ones((4, 4), np.float64)),
+        # int32 — the lowering is float-only (f32/f64/f16/bf16); integer dtypes
+        # are outside the boundary table (f64 is now wired, see test below).
+        (np.ones((4, 4), np.int32), np.ones((4, 4), np.int32)),
         # Shape mismatch — elementwise requires equal shapes.
         (np.ones((4, 8), np.float32), np.ones((8, 4), np.float32)),
         # Scalar (rank-0) — Phase 1 boundary requires rank >= 1.
